Remove static Agent.Current
