--- src/Elastic.OpenTelemetry/Diagnostics/DiagnosticEvent.cs
@@ -1,21 +1,0 @@
-// Licensed to Elasticsearch B.V under one or more agreements.
-// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
-// See the LICENSE file in the project root for more information
-using System.Diagnostics;
-
-namespace Elastic.OpenTelemetry.Diagnostics;
-
-internal class DiagnosticEvent(Activity? activity = null)
-{
-	public int ManagedThreadId { get; } = Environment.CurrentManagedThreadId;
-
-	public DateTime DateTime { get; } = DateTime.UtcNow;
-
-	public Activity? Activity { get; } = activity ?? Activity.Current;
-}
-
-internal class DiagnosticEvent<T>(T data, Activity? activity = null) : DiagnosticEvent(activity)
-{
-	public T Data { get; init; } = data;
-}
-
